Code refactor for TIM generator
